More thing and weapon tables for ID24

+++ src/main/java/net/mtrop/doom/tools/decohack/contexts/AbstractPatchContext.java
@@ -808,7 +808,7 @@ public abstract class AbstractPatchContext<P extends DEHPatch> implements DEHPat
 	}
 
 	// Throws IndexOutOfBoundsException if out of range.
-	protected void checkIndexRange(int index, int count)
+	protected void checkIndexRange(int index, long count)
 	{
 		if (index < 0 || index >= count)
 			throw new IndexOutOfBoundsException("Index cannot be less than 0 or greater than " + (count - 1));
